+++ tla-import/src/main/scala/at/forsyte/apalache/io/json/JsonEncoder.scala
@@ -2,6 +2,10 @@ package at.forsyte.apalache.io.json
 
 import at.forsyte.apalache.tla.lir.{TlaDecl, TlaEx, TlaModule}
 
+/**
+ * A JsonEncoder defines a conversion from a TLA+ expression to a json (as represented by T)
+ * @tparam T Any class extending JsonRepresentation
+ */
 trait JsonEncoder[T <: JsonRepresentation] {
   def apply(ex: TlaEx): T
   def apply(decl: TlaDecl): T
