+++ tla-import/src/main/scala/at/forsyte/apalache/io/json/JsonRepresentation.scala
@@ -1,9 +1,9 @@
 package at.forsyte.apalache.io.json
 
+/**
+ * A representation of json. The concrete implementation may depend on external json libraries.
+ * Defines a toString method, to be used when performing IO.
+ */
 trait JsonRepresentation {
   def toString: String
-}
-
-sealed case class UJsonRep(protected[json] val value: ujson.Value) extends JsonRepresentation {
-  override def toString: String = ujson.write(value, indent = 2, escapeUnicode = false)
 }
