+++ tla-import/src/main/scala/at/forsyte/apalache/io/json/impl/TlaToUJson.scala
@@ -1,0 +1,8 @@
+package at.forsyte.apalache.io.json.impl
+
+import at.forsyte.apalache.io.json.TlaToJson
+
+/**
+ * A json encoder, using the UJson representation
+ */
+object TlaToUJson extends TlaToJson[UJsonRep](UJsonFactory)
